+++ test/milestones.test.ts
@@ -52,7 +52,7 @@ function setup(team: Team): { container: HTMLElement; store: Store; pm: PaneMana
 
 function render(container: HTMLElement, loc: Loc, store: Store, pm: PaneManager, paneIdx: 0 | 1 = 0): void {
   const searchIndex = createSearchIndex(() => store.doc, () => store.rev)
-  const ctx: ModuleCtx = { store, pm, paneIdx, locale: 'en-US', searchIndex }
+  const ctx: ModuleCtx = { store, pm, paneIdx, locale: 'en-US', searchIndex, saveStatus: { requestSaveNow: () => {}, subscribeSaveState: () => () => {} } }
   renderMilestones(container, loc, ctx)
 }